+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fetch conversation container item if integrity proof is not present.
 - De-duplicate media attachments by URL.
 - Use primary gateway address for filtering messages from portable actors.
+- Ignore images in `icon` field if object type is not `Video`.
 
 ### Fixed